+++ py42/_internal/clients/security.py
@@ -3,11 +3,30 @@ from py42._internal.base_classes import BaseAuthorityClient
 
 
 def get_normalized_security_event_plan_info(security_client, user_uid, **kwargs):
-    # Raises HTTPError if no plans are found.
+    """Returns a dictionary of plan UIDs to their locations.
+
+    Raises:
+        HTTPError if no plans are found.
+
+    Note: The data model changed with Code42 6.7.
+
+    Example return value (6.7 and later):
+
+        {"111111111111111111": [{"destinationGuid": "4", "nodeGuid": "41"},
+                                {"destinationGuid": "5", "nodeGuid": "51"}],
+         "222222222222222222": [{"destinationGuid": "4", "nodeGuid": "41"},
+                                {"destinationGuid": "5", "nodeGuid": "51"}]}
+
+    Pre Code42 6.7:
+
+        {"111111111111111111": [{"destinationGuid": "4"},
+                                {"destinationGuid": "5"}]}
+    """
     # Normally we would just be able to call the SecurityEventLocations endpoint and use what comes back,
     # but due to this endpoint having its format completely change across C42 versions, we need to perform
     # transformations on the response in order to standardize.
     # TODO: once we have proper types this should be refactored to not use so many magical strings
+    archive_locations_by_plan = {}
     response = security_client.get_security_event_locations(user_uid, force_sync=True, **kwargs)
     if response is not None:
         # Prior to Code42 6.7, the SecurityEventLocation api returned data back in a different format,
@@ -15,16 +34,21 @@ def get_normalized_security_event_plan_info(security_client, user_uid, **kwargs)
         destination_list = util.get_obj_from_response(response, "storageNodeGuids")
         if destination_list:
             plan_uid = util.get_obj_from_response(response, "planUid")
-            plan_dict = {plan_uid: [{"destinationGuid": destination}] for destination in destination_list}
+            archive_locations_by_plan = {
+                plan_uid: [{"destinationGuid": destination}] for destination in destination_list
+            }
         else:
-            location_list = util.get_obj_from_response(response, "securityPlanLocationsByDestination")
-            plan_dict = {z: [{"destinationGuid": x["destinationGuid"], "nodeGuid": y["nodeGuid"]}]
-                         for x in location_list
-                         for y in x["securityPlanLocationsByNode"]
-                         for z in y["securityPlanUids"]}
-
-        return plan_dict
-    return {}
+            destinations = util.get_obj_from_response(response, "securityPlanLocationsByDestination")
+            for destination in destinations:
+                for node in destination["securityPlanLocationsByNode"]:
+                    for plan_uid in node["securityPlanUids"]:
+                        plan_locations = archive_locations_by_plan.get(plan_uid)
+                        if plan_locations is None:
+                            archive_locations_by_plan[plan_uid] = []
+                            plan_locations = archive_locations_by_plan.get(plan_uid)
+                        plan_locations.append({"destinationGuid": destination["destinationGuid"],
+                                               "nodeGuid": node["nodeGuid"]})
+    return archive_locations_by_plan
 
 
 class SecurityClient(BaseAuthorityClient):