+++ CHANGELOG.md
@@ -12,6 +12,9 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ### Fixed
 - Bug in authentication handling logic that caused authentication tokens to not automatically renew properly when
  they expired.
+- Bug in creating security plan clients that caused some clients to not be created for users with multiple plans or
+ archives
+
 
 
 ## 0.1.6 – 2019-07-30